Rename/move to `plugins.MetadataSourcePlugin`, fix formatting
