+++ docs/changelog.rst
@@ -66,10 +66,9 @@ New features:
 * The 'data_source' field is now also applied as an album-level flexible
   attribute during imports, allowing for more refined album level searches.
   :bug:`3350` :bug:`1693`
-* :doc:`/plugins/deezer`:
-   * Added Deezer plugin as an import metadata provider: you can now match tracks
-     and albums using the `Deezer`_ database.
-     Thanks to :user:`rhlahuja`.
+* :doc:`/plugins/deezer`: Added Deezer plugin as an import metadata provider:
+  you can now match tracks and albums using the `Deezer`_ database.
+  Thanks to :user:`rhlahuja`.
 
 Fixes:
 
@@ -128,6 +127,10 @@ For plugin developers:
   longer separate R128 backend instances. Instead the targetlevel is passed to
   ``compute_album_gain`` and ``compute_track_gain``.
   :bug:`3065`
+* The ``beets.plugins.MetadataSourcePlugin`` base class has been added to
+  simplify development of plugins which query album, track, and search
+  APIs to provide metadata matches for the importer. Refer to the Spotify and
+  Deezer plugins for examples of using this template class.
 
 For packagers:
